fix(storage): break circular dependency storage<->kernel (issue #312)
SnapshotManager.ts imported type ProcessEntry from kernel/ProcessTable,
closing a package cycle with kernel's value imports of SnapshotManager.

Move ProcessStatus/ProcessStatusValue/ProcessEntry to
packages/shared/types.ts (the project's mandated home for shared data
shapes), re-export them from ProcessTable so existing kernel importers
stay unchanged, and switch SnapshotManager to import from shared.

Verified: typecheck (web+cli), 141/141 tests, self-analysis
detectRepositoryPattern findings: 0 (no package-level cycles left).
